// src/lib/model/CollItem.ts
export default interface CollItem {
  added: string;
  album_id: number;
  album_title: string;
  // also_collected_count: 189
  band_id: number;
  // band_image_id: null
  // band_location: null
  band_name: string;
  band_url: string;
  currency: string;
  // discount: null
  fan_id: number;
  // featured_track: 3672987069
  // featured_track_duration: 826.56
  // featured_track_encodings_id: 1120015807
  // featured_track_is_custom: false
  // featured_track_license_id: null
  // featured_track_number: 1
  // featured_track_title: "O World! I Remain No Longer Here."
  // featured_track_url: null
  // genre_id: 18
  // gift_id: null
  // gift_recipient_name: null
  // gift_sender_name: null
  // gift_sender_note: null
  // has_digital_download: null
  // hidden: null
  // index: null
  // is_giftable: true
  // is_preorder: false
  // is_private: false
  // is_purchasable: true
  // is_set_price: false
  // is_subscriber_only: false
  // is_subscription_item: false
  // item_art: {thumb_url: "https://f4.bcbits.com/img/a1955602981_3.jpg", art_id: 1955602981,…}
  item_art_id: number;
  // item_art_ids: null
  item_art_url: string;
  item_id: number;
  item_title: string;
  item_type: string;
  item_url: string;
  // label: null
  // label_id: null
  // licensed_item: null
  listen_in_app_url: string;
  // merch_ids: [310904532, 2394825112, 2499238452, 3242151506]
  // merch_snapshot: null
  // merch_sold_out: false
  // message_count: null
  // num_streamable_tracks: 4
  // package_details: null
  // price: 0
  purchased: string;
  // release_count: null
  // releases: null
  // require_email: null
  // sale_item_id: 107060402
  // sale_item_type: "p"
  // service_name: null
  // service_url_fragment: null
  token: string;
  tralbum_id: number;
  tralbum_type: string;
  updated: string; // date
  release_date: Date;
  // url_hints: {item_type: "a", custom_domain_verified: null, slug: "no-light-ever", subdomain: "glacierma",…}
  // variant_id: null
  // why: null
}